% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_dataframe.R
\name{add_to_dataframe}
\alias{add_to_dataframe}
\title{Join a dataframe and a named vector}
\usage{
add_to_dataframe(dat, vec, by, newcol = "val")
}
\arguments{
\item{dat}{dataframe}

\item{vec}{named vector whose names are a column of the dataframe}

\item{by}{the name of the column that joining should be performed on.}

\item{newcol}{the name of the new column to add from vec. Default: 'val'}
}
\value{
dataframe with the same number of rows as 'dat' and an additional column corresponding to the values of 'vec'.
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
\code{\link[dplyr]{join}}
}
