% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggr_dup_genes.R
\name{aggr_dup_genes}
\alias{aggr_dup_genes}
\title{Aggregate duplicate gene rows}
\usage{
aggr_dup_genes(m, stat = c("sum", "mean", "max", "median"), na.rm = T)
}
\arguments{
\item{m}{matrix with genes as rows}

\item{stat}{statistic to aggregate duplicate gene rows. Defaults to 'sum'. Default: c("sum", "mean", "max", "median")}
}
\value{
matrix with unique gene rows
}
\description{
Aggregate duplicate gene rows. Duplicate rows are collapsed into one by taking the colSums, colMeans, colMaxs or colMedians.
}
\seealso{
\code{\link[matrixStats]{rowSums2}},\code{\link[matrixStats]{rowMeans2}},\code{\link[matrixStats]{rowRanges}},\code{\link[matrixStats]{rowMedians}}
}
