% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annomaps.R
\name{annomaps}
\alias{annomaps}
\title{many-bar annotations ggplot}
\usage{
annomaps(..., titles = NULL, flip = F, ratio = 0.03, mar = 0.015,
  angle = T)
}
\arguments{
\item{...}{named character vectors of observations and groups (names). All vectors must contain the same observations, and the ordering in the first vector provided is preserved.}

\item{titles}{character vector of annomap titles. Default: NULL}

\item{flip}{observations along y axis instead of x? Default: F}

\item{ratio}{aspect ratio. If flip = T, ratio = 1/ratio.. Default: 0.03}

\item{mar}{see plot.margin in ggplot2::themes for details. Default: 0.015}

\item{angle}{logical indicating whether to angle title if flip = T. Default: T}
}
\value{
ggplot object consisting of ggplots combined.
}
\description{
many-bar annotations ggplot
}
\seealso{
\code{\link[ggplot2]{waiver}},\code{\link[ggplot2]{theme}}
}
