% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_bulk.R
\name{as_bulk}
\alias{as_bulk}
\title{Convert sc-matrix to simulated bulk}
\usage{
as_bulk(m, groups = NULL, isLog = TRUE, by = "mean")
}
\arguments{
\item{m}{non-row-centered logTPM or TPM matrix.}

\item{groups}{if not NULL, return a simulated bulk column per group in groups. Default: NULL}

\item{isLog}{set to FALSE if 'm' is in TPM form. Default: TRUE}
}
\value{
simulated bulk matrix. If groups is NULL, one column is returned for the average across all cells. Else one column per group in 'groups'.
}
\description{
Convert sc-matrix to simulated bulk
}
