% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_bulk.R
\name{as_bulk_samples}
\alias{as_bulk_samples}
\title{Convert sc-matrix to simulated bulk samples matrix}
\usage{
as_bulk_samples(
  m,
  samples = split_by_sample_names(colnames(m)),
  isLog = TRUE,
  by = "median"
)
}
\arguments{
\item{m}{non-row-centered logTPM or TPM matrix.}

\item{samples}{a list of cell IDs per sample. Default: split_by_sample_names(colnames(m))}

\item{isLog}{set to FALSE if 'm' is in TPM form. Default: TRUE}
}
\value{
simulated bulk matrix,with as many columns as there are samples in the <samples> list.
}
\description{
Convert sc-matrix to simulated bulk samples matrix
}
