% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permuteSigScores.R
\name{binom.test_matrix}
\alias{binom.test_matrix}
\title{Binomial test on matrix}
\usage{
binom.test_matrix(m, adjust.method = "fdr", MARGIN = 1)
}
\arguments{
\item{m}{matrix (logical or binary)}

\item{adjust.method}{p-value adjustment method. Default: 'fdr'}

\item{MARGIN}{1 for rows, 2 for columns. Default: 1}
}
\value{
adjusted p-values
}
\description{
Performs binomial test on each row (or column) of a logical/binary matrix to test if there are more TRUE values than expected by chance. Tests the alternative hypothesis that the proportion of successes is greater than expected.
}
