% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_bulk_states.R
\name{bootstrap_summary}
\alias{bootstrap_summary}
\title{Summary Statistics for Bootstrapped State Fractions}
\usage{
bootstrap_summary(bootstraps)
}
\arguments{
\item{bootstraps}{a matrix or dataframe of data values in rows by bootstrap repeats in columns}
}
\value{
dataframe with the mean and SEM across all bootstrap repeats (columns) for each row in <bootstraps>
}
\description{
Compute summary statistics for bootstrapped data
}
\seealso{
\code{\link[matrixStats]{rowMads}}
}
