% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.methods.R
\docType{data}
\name{cluster.methods}
\alias{cluster.methods}
\title{agglomeration method in hierarchical clustering
agglomeration method in hierarchical clustering when grouping members into a tree structure.}
\format{
cluster.methods
}
\usage{
cluster.methods
}
\description{
agglomeration method in hierarchical clustering
agglomeration method in hierarchical clustering when grouping members into a tree structure.
}
\keyword{datasets}
