% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_exclude.R
\name{col_exclude}
\alias{col_exclude}
\title{Drop columns whose names match a pattern}
\usage{
col_exclude(m, pattern)
}
\arguments{
\item{m}{Input matrix or data frame.}

\item{pattern}{Pattern to look for. The default interpretation is a regular expression, as described in stringi::stringi-search-regex. Control options with 'regex()'.}
}
\value{
A matrix.
}
\description{
col_exclude() is a wrapper around m[, !str_detect(colnames(m), pattern)]. Equivalent to col_subset() with negate = TRUE.
}
\details{
Vectorised over colnames('m') and 'pattern'
}
