% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_subset.R
\name{col_subset}
\alias{col_subset}
\title{Keep columns whose names match a pattern}
\usage{
col_subset(m, pattern, negate = FALSE)
}
\arguments{
\item{m}{Input matrix or data frame.}

\item{pattern}{Pattern to look for. The default interpretation is a regular expression, as described in stringi::stringi-search-regex. Control options with 'regex()'.}

\item{negate}{If 'TRUE', return non-matching columns.}
}
\value{
A matrix.
}
\description{
col_subset() is a wrapper around m[, str_detect(colnames(m), pattern)]
}
\details{
Vectorised over colnames('m') and 'pattern'
}
