% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-ops.R
\name{colcenter}
\alias{colcenter}
\title{Center a matrix column-wise}
\usage{
colcenter(m, by = "mean")
}
\arguments{
\item{m}{a matrix or Matrix}

\item{by}{either "mean", "median" or a numeric vector of length equal to the number of columns of ‘m’. Default: "mean"}
}
\value{
column-centered matrix
}
\description{
Center a matrix column-wise
}
