% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldetected.R
\name{coldetected}
\alias{coldetected}
\title{Number of non-zero values per column}
\usage{
coldetected(m)
}
\arguments{
\item{m}{matrix}
}
\value{
numeric vector
}
\description{
Number of non-zero values per column
}
\seealso{
\code{\link[matrixStats]{rowCounts}}
 \code{\link[stats]{setNames}}
}
