% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_group2_dea.R
\name{combine_group2_dea}
\alias{combine_group2_dea}
\title{Aggregate `scalop::dea` output for one group against many in turn}
\usage{
combine_group2_dea(L.dea, coverage = 0.5, return.avg.lfc = FALSE)
}
\arguments{
\item{L.dea}{list of gene-named-numeric log2-fold change values}

\item{coverage}{prioritise genes that appear as DE in at least this fraction of the dea result vectors in 'L.dea'. Default: 0.5}

\item{return.avg.lfc}{logical; if FALSE (default) return gene names only (ordered by average log2 fold change). If TRUE, return average values. Default: FALSE}
}
\description{
Aggregate `scalop::dea` output for one group against many in turn
}
