% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.methods.R
\docType{data}
\name{cor.methods}
\alias{cor.methods}
\title{Correlation method
method for computing correlation coefficients}
\format{
cor.methods
}
\usage{
cor.methods
}
\description{
Correlation method
method for computing correlation coefficients
}
\keyword{datasets}
