% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm.R
\name{cpm}
\alias{cpm}
\title{Convert Counts to Counts-Per-Million (CPM)}
\usage{
cpm(counts)
}
\arguments{
\item{counts}{matrix of raw counts (gene rows; cell columns)}
}
\value{
scaled matrix of counts-per-million values
}
\description{
Convert counts to CPM, by scaling raw counts such that they sum to 1 million per cell.
}
