% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crorder.R
\name{crorder}
\alias{crorder}
\title{Ordered correlation matrix}
\usage{
crorder(m, col = T, row = T)
}
\arguments{
\item{m}{a matrix}

\item{col}{logical indicating whether to reorder columns. Default: T}

\item{row}{logical indicating whether to reorder rows. Default: T}
}
\value{
ordered correlation matrix.
}
\description{
Computes correlation matrix and reorders its columns and rows by hierarchical clustering of the correlation matrix.
}
