% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea}
\alias{dea}
\title{Differential Expression Analysis}
\usage{
dea(
  m,
  groups,
  is.log = T,
  fc = 2L,
  p = 0.05,
  p.adjust.method = "BH",
  sortby = "fc",
  val = NULL
)
}
\arguments{
\item{m}{numeric matrix. The matrix should not contain NA values.}

\item{groups}{the group(s) to be tested. Each group should be a character vector or a two-level factor in which the first corresponds to the group. Multiple groups must be supplied as a list.}

\item{is.log}{is the data in log2 space? Fold change cutoff and calculation are adjusted accordingly. Default: T}

\item{fc}{fold change cutoff; numeric value or NULL. Default: 2}

\item{p}{p-value cutoff; numeric value between 0 and 1 or NULL. Default: 0.05}

\item{p.adjust.method}{correction method for multiple testing or 'none' if not desired. See stats::p.adjust.methods for options. Default: 'BH'}

\item{sortby}{sort genes by 'fc', 'p' or NULL. Default: 'fc'}

\item{val}{return a named numeric vector of 'fc' or 'p' values. If NULL, returns the full dataframe. Default: NULL}
}
\value{
a numeric vector of gene fold changes or p-values or a dataframe with both.
}
\description{
Differential Expression Analysis. Compute gene-level fold changes and (adjusted) p-values between cell clusters of an expression matrix.
}
\seealso{
\code{\link[stats]{character(0)}}
}
