% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_colours-data.R
\docType{data}
\name{discrete_colours}
\alias{discrete_colours}
\title{The 297 discrete colours available across different RColorBrewer palettes.}
\format{
A character vector.
}
\usage{
data(discrete_colours)
}
\description{
The first 24 were selected and ordered carefully (with qualitative data in mind) and the rest was taken and shuffled.
}
\examples{
data(discrete_colours)
scales::show_col(discrete_colours)
}
\keyword{datasets}
