% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.methods.R
\docType{data}
\name{dist.methods}
\alias{dist.methods}
\title{distance metric
method for computing distances from a matrix object}
\format{
dist.methods
}
\usage{
dist.methods
}
\description{
distance metric
method for computing distances from a matrix object
}
\keyword{datasets}
