% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl2symbol.R
\docType{data}
\name{ensembl2symbol}
\alias{ensembl2symbol}
\title{ensembl2symbol}
\format{
A character vector.
}
\usage{
symbols = ensembl2symbol[ensembl_gene_ids]
}
\description{
ensembl2symbol
}
\keyword{datasets}
