% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_genes.R
\name{entrez2ucsc}
\alias{entrez2ucsc}
\title{FUNCTION_TITLE}
\usage{
entrez2ucsc(keys = NULL, multiVals = "first")
}
\arguments{
\item{keys}{PARAM_DESCRIPTION, Default: NULL}

\item{multiVals}{PARAM_DESCRIPTION, Default: 'first'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[Homo.sapiens]{Homo.sapiens}}
 \code{\link[AnnotationDbi]{AnnotationDb-objects}}
}
