% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprs_levels.R
\name{exprs_levels}
\alias{exprs_levels}
\title{Convert normalised counts into log2 counts}
\usage{
exprs_levels(m, bulk = F, log_scale = F, unlog = F)
}
\arguments{
\item{m}{matrix of counts}

\item{bulk}{divides the counts by 10 to estimate the complexity of single cell libraries in the order of 100,000 transcripts to avoid counting each transcript ~10 times, which may inflate the difference between the expression level of a gene in cells in which the gene is detected and those in which it is not detected.  Default: FALSE.}

\item{log_scale}{should the counts be placed on a log2 scale.}

\item{unlog}{does the reverse of log_scale.}
}
\value{
a matrix
}
\description{
Logs and un-Logs a normalised counts matrix. Also accounts for the complexity of single cell libraries transcripts.
}
