% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sample_names.R
\name{extract_sample_names}
\alias{extract_sample_names}
\title{Extract Sample Names}
\usage{
extract_sample_names(x, ..., sample_names = NULL)
}
\arguments{
\item{x}{character vector of strings to extract sample names from}

\item{...}{other arguments passed to `scalop::unique_sample_names`. Relevant only if sample_names is not provided.}

\item{sample_names}{character vector of unique sample names to extract from x. Default: NULL}
}
\value{
character vector of sample names of the same length as x.
}
\description{
Convenient wrapper function for `scalop::extract_substring`. Extract sample names from character vector. Sample names can be provided or, if not, then are found with `scalop::unique_sample_names`.
}
