% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_substring.R
\name{extract_substring}
\alias{extract_substring}
\title{Find and Extract substrings}
\usage{
extract_substring(x, choose)
}
\arguments{
\item{x}{character vector of strings to extract substrings from.}

\item{choose}{character vector of unique substrings of which one is expected to match a given string in the character vector x. Default: NULL}
}
\value{
character vector of detected substrings of the same length as x.
}
\description{
Detect substring and extract if present for each string in a character vector.
}
\details{
if there is more than one match, the longest substring is taken. i.e. if HEL and HELLO are in substring vector choose, HELLO will be returned.
}
\seealso{
\code{\link[stringr]{str_extract}}
}
