% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigScores.R
\name{filter_groups}
\alias{filter_groups}
\title{Filter genes in groups according to reference}
\usage{
filter_groups(groups, ref, conserved = 0.7)
}
\arguments{
\item{groups}{a character vector of genes or list of character vectors to filter}

\item{ref}{reference genes to filter groups according to}

\item{conserved}{the minimum allowed fraction of genes retained after filtering. Default: 0.7}
}
\value{
a filtered list of groups. Returns NULL if no genes in any groups are left. Groups with less than <conserved> fraction of genes retained will be removed. Returns input <groups> if no genes are missing.
}
\description{
Filter genes in groups according to reference
}
