% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_signatures.R
\name{filter_signatures}
\alias{filter_signatures}
\title{Filter Out Lowly-Correlated Genes from Signature(s)}
\usage{
filter_signatures(m, sigs, filter.threshold = 0.4)
}
\arguments{
\item{m}{expression matrix of genes X cells to score. Not centered.}

\item{sigs}{list of gene signatures to be refined.}

\item{filter.threshold}{minimum pearson's r below which genes are filtered out of the signature. Default: 0.4}
}
\value{
a filtered list of signatures, whose genes' expression levels all had correlation values to the signature score above or equal to <filter.threshold>.
}
\description{
This function removes signature genes whose expression level is poorly correlated to the signature score. This is relevant if the signatures themselves were defined from a different dataset, as might be the case in applying single-cell signatures to bulk, or in moving between datasets generated with different protocols (e.g. single-nuclei RNA seq. vs single-cell RNA seq.).
}
