% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigScores.R
\name{filter_sigs}
\alias{filter_sigs}
\title{Filter genes in sigs according to reference}
\usage{
filter_sigs(sigs, ref, conserved = 0.7)
}
\arguments{
\item{sigs}{a character vector of genes or list of character vectors to filter}

\item{ref}{reference genes to filter sigs according to}

\item{conserved}{the minimum allowed fraction of genes retained after filtering. Default: 0.7}
}
\value{
a filtered list of sigs. Returns NULL if no genes in any sigs are left. sigs with less than <conserved> fraction of genes retained will be removed. Returns input <sigs> if no genes are missing.
}
\description{
Filter genes in sigs according to reference
}
