% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.R
\name{flip}
\alias{flip}
\title{Flip the (nested) elements of a character vector (or list) with its names}
\usage{
flip(X)
}
\arguments{
\item{X}{A named character vector or a named list of character vectors.}
}
\value{
A named character vector or a named list of character vectors. If the former, will be of the same length as <X>. If the latter, will be of the same length as there are unique elements across all vectors in the list.
}
\description{
A convenience function to switch the names and elements of a named character vector or a named list of character vectors.
}
\seealso{
\code{\link[stats]{setNames}}
}
