% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldchange.R
\name{foldchange}
\alias{foldchange}
\title{fold changes}
\usage{
foldchange(x, y, ...)
}
\arguments{
\item{x}{matrix or character vector with matrix column names}

\item{y}{matrix to compare against or matrix with x and y column names}

\item{...}{is.log and cutoff}
}
\value{
fold changes
}
\description{
fold changes
}
