% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futurama-data.R
\docType{data}
\name{futurama}
\alias{futurama}
\title{futurama, 12 Discrete Colours}
\format{A character vector.}
\usage{
data(futurama)
}
\description{
futurama, 12 Discrete Colours
}
\examples{
data(futurama)
scales::show_col(futurama)
}
\keyword{datasets}
