% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneLength.R
\name{geneLength}
\alias{geneLength}
\title{FUNCTION_TITLE}
\usage{
geneLength(genes = NULL)
}
\arguments{
\item{genes}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[Homo.sapiens]{Homo.sapiens}}
 \code{\link[AnnotationDbi]{AnnotationDb-objects}}
 \code{\link[dplyr]{filter}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise}},\code{\link[dplyr]{rename}}
}
