% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericModel.R
\name{genericModel}
\alias{genericModel}
\title{FUNCTION_TITLE}
\usage{
genericModel(mod, formula, data, ...)
}
\arguments{
\item{mod}{PARAM_DESCRIPTION}

\item{formula}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
