% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_genes.R
\name{get_genes}
\alias{get_genes}
\title{FUNCTION_TITLE}
\usage{
get_genes(keytype)
}
\arguments{
\item{keytype}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[Homo.sapiens]{Homo.sapiens}}
 \code{\link[AnnotationDbi]{AnnotationDb-objects}}
}
