% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signatures.R
\name{get_signatures}
\alias{get_signatures}
\title{FUNCTION_TITLE}
\usage{
get_signatures(pattern, signatures = NormalSignatures)
}
\arguments{
\item{pattern}{PARAM_DESCRIPTION}

\item{signatures}{PARAM_DESCRIPTION, Default: scalop::NormalSignatures}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[stringr]{str_detect}}
}
