% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snn_graph_clusters.R
\name{get_snn_graph_clusters}
\alias{get_snn_graph_clusters}
\title{FUNCTION_TITLE}
\usage{
get_snn_graph_clusters(m, k = 10, d = nrow(m), return.list = FALSE)
}
\arguments{
\item{m}{PARAM_DESCRIPTION}

\item{k}{PARAM_DESCRIPTION, Default: 10}

\item{d}{PARAM_DESCRIPTION, Default: nrow(m)}

\item{return.list}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[scran]{buildSNNGraph}}
 \code{\link[igraph]{cluster_walktrap}}
}
