% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_genes.R
\name{get_ucsc}
\alias{get_ucsc}
\title{FUNCTION_TITLE}
\usage{
get_ucsc()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[Homo.sapiens]{Homo.sapiens}}
 \code{\link[AnnotationDbi]{AnnotationDb-objects}}
}
