% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGRCh.R
\name{getGRCh}
\alias{getGRCh}
\title{Get hg19 or hg38 information from biomaRt}
\usage{
getGRCh(genome = "hg19", as.granges = FALSE)
}
\arguments{
\item{genome}{hg38 or hg19}

\item{as.granges}{Output as GRanges or data.frame}
}
\description{
Get hg19 or hg38 information from biomaRt
}
