% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdensity.R
\name{ggDensity}
\alias{ggDensity}
\title{FUNCTION_TITLE}
\usage{
ggDensity(
  data,
  x,
  xlab = "Cell Score",
  mode.names = c("Mode1", "Mode2"),
  ...,
  prob = 0.95,
  coverage = 0.95,
  min.size = 10,
  na.action = c("show", "hide", "combine"),
  combine.with = 1
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION}

\item{xlab}{PARAM_DESCRIPTION, Default: 'Cell Score'}

\item{mode.names}{PARAM_DESCRIPTION, Default: c("Mode1", "Mode2")}

\item{...}{PARAM_DESCRIPTION}

\item{prob}{PARAM_DESCRIPTION, Default: 0.95}

\item{coverage}{PARAM_DESCRIPTION, Default: 0.95}

\item{min.size}{PARAM_DESCRIPTION, Default: 10}

\item{na.action}{PARAM_DESCRIPTION, Default: c("show", "hide", "combine")}

\item{combine.with}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{setNames}}
 \code{\link[ggpubr]{ggdensity}}
 \code{\link[dplyr]{filter}},\code{\link[dplyr]{mutate}}
}
