% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggoutliers.R
\name{ggoutliers}
\alias{ggoutliers}
\title{Plot model and Outliers}
\usage{
ggoutliers(
  formula,
  data,
  mod = loess,
  n.sd = 1.5,
  ...,
  selected = NULL,
  add.to.selected = NULL,
  ngenes = 6,
  yvar.ngenes = ngenes,
  xvar.ngenes = ngenes,
  gene.min = 0.5,
  alt.min = 0.5,
  xlab.min = 0,
  ylab.min = 0,
  xlab.max = NULL,
  ylab.max = NULL,
  text.size = 16,
  legend.text.size = 12,
  label.size = 3.5,
  segment.colour = "grey85",
  segment.size = 0.15,
  min.segment.length = 0.5,
  nudge_y = 0,
  force = 5,
  xlab = all.vars(formula)[2],
  ylab = all.vars(formula)[1],
  selected.col = "red",
  subtitle = NULL,
  midline = TRUE,
  outerlines = FALSE,
  title = NULL,
  caption = NULL
)
}
\arguments{
\item{formula}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION}

\item{mod}{PARAM_DESCRIPTION, Default: loess}

\item{n.sd}{PARAM_DESCRIPTION, Default: 1.5}

\item{...}{PARAM_DESCRIPTION}

\item{selected}{PARAM_DESCRIPTION, Default: NULL}

\item{add.to.selected}{PARAM_DESCRIPTION, Default: NULL}

\item{ngenes}{PARAM_DESCRIPTION, Default: 7}

\item{yvar.ngenes}{PARAM_DESCRIPTION, Default: ngenes}

\item{xvar.ngenes}{PARAM_DESCRIPTION, Default: ngenes}

\item{gene.min}{PARAM_DESCRIPTION, Default: 3}

\item{alt.min}{PARAM_DESCRIPTION, Default: 1}

\item{xlab.min}{PARAM_DESCRIPTION, Default: 0}

\item{ylab.min}{PARAM_DESCRIPTION, Default: 0}

\item{text.size}{PARAM_DESCRIPTION, Default: 14}

\item{legend.text.size}{PARAM_DESCRIPTION, Default: 12}

\item{label.size}{PARAM_DESCRIPTION, Default: 5}

\item{force}{PARAM_DESCRIPTION, Default: 20}

\item{xlab}{PARAM_DESCRIPTION, Default: all.vars(formula)[2]}

\item{ylab}{PARAM_DESCRIPTION, Default: all.vars(formula)[1]}

\item{selected.col}{PARAM_DESCRIPTION, Default: 'red'}

\item{subtitle}{PARAM_DESCRIPTION, Default: NULL}

\item{title}{PARAM_DESCRIPTION, Default: NULL}

\item{caption}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Plot model and Outliers
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{pull}}
 \code{\link[stats]{setNames}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{geom_path}},\code{\link[ggplot2]{scale_continuous}},\code{\link[ggplot2]{ggtheme}},\code{\link[ggplot2]{scale_manual}},\code{\link[ggplot2]{theme}}
 \code{\link[ggrepel]{geom_text_repel}}
}
