% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpubr-data.R
\docType{data}
\name{ggpubr}
\alias{ggpubr}
\title{ggpubr, 3 Discrete Colours}
\format{
A character vector.
}
\usage{
data(ggpubr)
}
\description{
ggpubr, 3 Discrete Colours
}
\examples{
data(ggpubr)
scales::show_col(ggpubr)
}
\keyword{datasets}
