% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvolcano.R
\name{ggvolcano}
\alias{ggvolcano}
\title{FUNCTION_TITLE}
\usage{
ggvolcano(
  m = NULL,
  group = NULL,
  group2 = NULL,
  dea.res = NULL,
  selected = NULL,
  add.to.selected = NULL,
  selected.col = "red",
  alternative = NULL,
  xlab = "Log2 Fold Change",
  ylab = "- Log10 P-value (adjusted)",
  pmethod = "BH",
  ngenes = 10,
  text.size = 14,
  label.size = 5,
  symmetric = TRUE,
  add.line = TRUE,
  ymax = NULL,
  xlim = NULL,
  legend.text.size = 12
)
}
\arguments{
\item{m}{PARAM_DESCRIPTION, Default: NULL}

\item{group}{PARAM_DESCRIPTION, Default: NULL}

\item{group2}{PARAM_DESCRIPTION, Default: NULL}

\item{dea.res}{PARAM_DESCRIPTION, Default: NULL}

\item{selected}{PARAM_DESCRIPTION, Default: NULL}

\item{add.to.selected}{PARAM_DESCRIPTION, Default: NULL}

\item{selected.col}{PARAM_DESCRIPTION, Default: 'red'}

\item{alternative}{PARAM_DESCRIPTION, Default: NULL}

\item{xlab}{PARAM_DESCRIPTION, Default: 'Log2 Fold Change'}

\item{ylab}{PARAM_DESCRIPTION, Default: '- Log10 P-value (adjusted)'}

\item{pmethod}{PARAM_DESCRIPTION, Default: 'BH'}

\item{ngenes}{PARAM_DESCRIPTION, Default: 10}

\item{text.size}{PARAM_DESCRIPTION, Default: 14}

\item{label.size}{PARAM_DESCRIPTION, Default: 5}

\item{symmetric}{PARAM_DESCRIPTION, Default: TRUE}

\item{add.line}{PARAM_DESCRIPTION, Default: TRUE}

\item{ymax}{PARAM_DESCRIPTION, Default: NULL}

\item{xlim}{PARAM_DESCRIPTION, Default: NULL}

\item{legend.text.size}{PARAM_DESCRIPTION, Default: 12}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{pull}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{scale_continuous}},\code{\link[ggplot2]{ggtheme}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{scale_manual}},\code{\link[ggplot2]{geom_abline}}
 \code{\link[ggrepel]{geom_label_repel}}
}
