% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graster.R
\name{graster}
\alias{graster}
\title{Plot A Heatmap with ggplot2}
\usage{
graster(
  dat,
  x = Var1,
  y = Var2,
  fill = value,
  raster.labels = FALSE,
  limits = c(-0.5, 0.5),
  lim.find = F,
  lim.sym = T,
  x.name = NULL,
  y.name = NULL,
  angle = NULL,
  axis.rel = 1,
  title.rel = 1.1,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  text.size = 12,
  ratio = NULL,
  cols = NULL,
  col = NULL,
  na.value = "white",
  legend.position = "bottom",
  legend.justification = "right",
  legend.margin = margin(0, 0.15, 0, 0, "cm"),
  plot.margin = margin(0, 0.15, 0, 0, "cm"),
  legend.height = 0.4,
  legend.width = 0.6,
  legend.rel = 0.8,
  legend.colour = "black",
  ticks.linewidth = 0.5,
  breaks = waiver(),
  labels = waiver(),
  x.breaks = waiver(),
  x.axis.position = "bottom",
  y.breaks = waiver(),
  x.labels = waiver(),
  y.labels = waiver(),
  num = T,
  y.num = num,
  x.num = num,
  label.col = "black",
  legend.breaks = NULL,
  legend.labels = NULL,
  legend.title = NULL,
  legend.title.position = "top",
  legend.title.rel = 0.8,
  expand = c(0, 0)
)
}
\description{
Plot A Heatmap with ggplot2
}
