% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_reorder.R
\name{grouped_reorder}
\alias{grouped_reorder}
\title{Grouped Reorder and Ungroup}
\usage{
grouped_reorder(
  m,
  groups,
  interorder = FALSE,
  intraorder = TRUE,
  cor.method = "pearson",
  dist.method = "euclidean",
  cluster.method = "average",
  Names = FALSE
)
}
\arguments{
\item{m}{matrix to be reordered}

\item{groups}{groups within which reordering should take place.}

\item{interorder}{if TRUE, group order itself is reordered such that groups with similar profiles are placed near one another. E.g. Samples with similar average CNA profiles. Default: FALSE}

\item{cor.method}{desired correlation metric. Default: 'pearson'}

\item{dist.method}{desired distance metric to be used on top of correlation matrix. Default: 'euclidean'}

\item{cluster.method}{desired agglomeration method. Default: 'average'}

\item{Names}{return the vector of ordered IDs instead of the reordered matrix. Default: FALSE}
}
\value{
reordered matrix (same dimensions as input) or a character vector of ordered column names if Names = T.
}
\description{
Keep desired groups together but reorder members within each group. E.g. If cells from multiple samples are being reordered, samples can be kept together and cells reordered within. Option to also reorder the order of groups or samples, such that groups with similar average profiles are placed next to one another.
}
