% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_rowcenter.R
\name{grouped_rowcenter}
\alias{grouped_rowcenter}
\title{Center matrix rows by subsets of columns}
\usage{
grouped_rowcenter(m, groups = split_by_sample_names(colnames(m)), by = "mean")
}
\arguments{
\item{m}{a matrix}

\item{groups}{list of character vectors specifying the column ids to center by.  Default: split_by_sample_names(colnames(m))}

\item{by}{how centering should be performed. One of 'mean' or 'median'. Default: 'mean'}
}
\value{
A row-centered matrix, centered within groups and with the order of cells within groups retained. The matrix has the same number of rows as <m> and as many columns as there are cell ids in groups.
}
\description{
Center matrix rows by subsets of columns
}
