% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-ops.R
\name{has_dim}
\alias{has_dim}
\title{Check DF dimensions}
\usage{
has_dim(x)
}
\arguments{
\item{x}{a dataframe of a single matrix}
}
\value{
logical for whether dims attributes are present.
}
\description{
Checks if x has a dim attribute
}
