% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-ops.R
\name{have_equal_dims}
\alias{have_equal_dims}
\title{Equal dimensions}
\usage{
have_equal_dims(m1, m2)
}
\arguments{
\item{m1}{A  matrix}

\item{m2}{A  matrix}
}
\value{
logical
}
\description{
test if two matrices have equal dimensions
}
