% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-ops.R
\name{have_equal_rownames}
\alias{have_equal_rownames}
\title{Same row names}
\usage{
have_equal_rownames(m1, m2)
}
\arguments{
\item{m1}{A  matrix}

\item{m2}{A  matrix}
}
\value{
logical
}
\description{
test if two matrices have the same row names
}
