% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca.R
\name{hca}
\alias{hca}
\alias{hca_cor}
\alias{hca_dist}
\alias{hca_tree}
\alias{hca_order}
\alias{hca_groups}
\alias{hca_reorder}
\title{Hierarchical clustering analysis}
\usage{
hca(
  x,
  cor.method = cor.methods,
  dist.method = dist.methods,
  cluster.method = cluster.methods,
  max.dist = 1,
  h = NULL,
  k = NULL,
  min.size = 5,
  max.size = 0.5
)

hca_cor(
  x,
  return.steps = F,
  reorder = T,
  reorder.col = reorder,
  reorder.row = reorder,
  ...
)

hca_dist(x, return.steps = F, ...)

hca_tree(x, return.steps = F, ...)

hca_order(x, return.steps = F, ...)

hca_groups(x, return.steps = F, ...)

hca_reorder(x, col = T, row = T, cor.force = F, ...)
}
\arguments{
\item{x}{a matrix (features X observations), an object of class `dist` or an object of class `hclust`.}

\item{cor.method}{option if <x> is a matrix. One of 'pearson' (default), 'kendall', 'spearman' or NULL (no correlation coefficents computed). Default: scalop::cor.methods}

\item{dist.method}{option if <x> is a matrix or correlation matrix. One of 'euclidean' (default), 'maximum', 'manhattan', 'canberra', 'binary', 'minkowski' or NULL (in which case stats::as.dist() will be used). Default: scalop::dist.methods}

\item{cluster.method}{one of "average" (default), "complete", "single", "ward.D", "ward.D2", "mcquitty", "median" or "centroid". Default: scalop::cluster.methods}

\item{max.dist}{maximum distance between observations. Default: 1}

\item{h}{height(s) at which to cut the tree. If NULL, h will be set to all tree heights. Default: NULL}

\item{k}{number of groups to return from tree. If <h> and <k> are both not NULL, <k> takes precedence. Default: NULL}

\item{min.size}{PARAM_DESCRIPTION, Default: 5}

\item{max.size}{PARAM_DESCRIPTION, Default: 0.5}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Hierarchical clustering analysis
}
\details{
DETAILS
}
