% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{headl}
\alias{headl}
\title{Left-hand head a matrix or dataframe}
\usage{
headl(m, i = 6L)
}
\arguments{
\item{m}{a matrix or dataframe}

\item{i}{number of columns to print}
}
\description{
Left-hand head a matrix or dataframe
}
