% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{headr}
\alias{headr}
\title{Right-hand head a matrix or dataframe}
\usage{
headr(m, i = 6L)
}
\arguments{
\item{m}{a matrix or dataframe}

\item{i}{number of columns to print}
}
\description{
Right-hand head a matrix or dataframe
}
