% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{hierarchy}
\alias{hierarchy}
\title{Cell state hierarchy coordinates}
\usage{
hierarchy(m, quadrants = NULL, log.scale = T)
}
\arguments{
\item{m}{matrix of cell state scores (cell rows; state columns)}

\item{quadrants}{option to find pmax of a subset of columns and use this as a state quadrant. pass argument in the form of a list of character vectors (column names), where list elements should start from bottom and go from left to right. Default: NULL}

\item{log.scale}{scale the hierarchy coordinates in log2 space. Default: T}
}
\value{
dataframe with X and Y coordinate values.
 \code{\link[dplyr]{mutate}}
}
\description{
Cell state hierarchy coordinates
}
