% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_span.R
\name{hms_span}
\alias{hms_span}
\title{Subtract POSIXt Class Objects in HH:MM:SS Format}
\usage{
hms_span(start, end)
}
\arguments{
\item{start}{old Sys.time()}

\item{end}{new Sys.time()}
}
\value{
subtracted time in hh:mm:ss format
}
\description{
subtract POSIXt class objects and return value in HH:MM:SS format.
}
\details{
source: https://stackoverflow.com/questions/32100133/print-the-time-a-script-has-been-running-in-r answer by nrussell
}
