% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacFilt.R
\name{jacFilt}
\alias{jacFilt}
\title{Filter List By Jaccard}
\usage{
jacFilt(x, threshold = 0.7, which = FALSE)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{threshold}{PARAM_DESCRIPTION, Default: 0.7}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
*********
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
