% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ldcast}
\alias{ldcast}
\title{Turn a list into a dataframe}
\usage{
ldcast(L)
}
\arguments{
\item{L}{a list}
}
\value{
a dataframe with as many rows as the longest vector in the list
}
\description{
Turn a list into a dataframe
}
