% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logcpm.R
\name{logcpm}
\alias{logcpm}
\title{Convert CPM to logCPM}
\usage{
logcpm(m, bulk = F)
}
\arguments{
\item{m}{matrix of CPM values (gene rows; cell columns)}

\item{bulk}{if bulk then instead uses log2(CPM + 1). i.e. no scaling. Default: F}
}
\value{
logCPM matrix
}
\description{
Convert CPM to logCPM, i.e. using log2(CPM/10 + 1).
}
\details{
CPM/10 is used for single cells since 100,000 is a more reasonable estimate than 1,000,000 for the number of RNA transcripts in a cell. 1,000,000 is reasonable estimate for bulk samples that contain multiple cells.
}
