% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logRowMeans.R
\name{logRowMeans}
\alias{logRowMeans}
\title{Row Means of Log2 data}
\usage{
logRowMeans(x, type = c("tpm", "cpm"))
}
\arguments{
\item{x}{numeric matrix}

\item{type}{how was the data normalised? "tpm"; Transcripts Per Million. "cpm"; Counts Per Million. Default: "tpm"}
}
\value{
vector of log2 row means
}
\description{
calculate row means in normal space and then convert back to log2.
}
