% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_genes.R
\name{map_symbol}
\alias{map_symbol}
\title{FUNCTION_TITLE}
\usage{
map_symbol(
  keys = NULL,
  columns = NULL,
  add.columns = NULL,
  remove.columns = NULL
)
}
\arguments{
\item{keys}{PARAM_DESCRIPTION, Default: NULL}

\item{columns}{PARAM_DESCRIPTION, Default: NULL}

\item{add.columns}{PARAM_DESCRIPTION, Default: NULL}

\item{remove.columns}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[Homo.sapiens]{Homo.sapiens}}
 \code{\link[AnnotationDbi]{AnnotationDb-objects}}
}
