% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Markers_Normal.R
\docType{data}
\name{Markers_Normal}
\alias{Markers_Normal}
\title{List of Marker Genes for Normal Cell Types}
\format{A list of gene names  by cell type}
\usage{
Markers_Normal
}
\description{
List of Marker Genes for Normal Cell Types
}
\keyword{datasets}
