% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat01.R
\name{mat01}
\alias{mat01}
\title{Character List to Binary (membership) Matrix}
\usage{
mat01(L)
}
\arguments{
\item{L}{list of character vectors}
}
\value{
matrix of 0s and 1s with as many columns as there are vectors in the list and as many rows as there are unique elements across all vectors.
}
\description{
Convert a list of character vectors to a matrix with i columns and j rows, where i are the list vector names and j are the unique elements found across i vectors. Cell i,j has a value of 1 if element j exists in vector i, and a value of 0 otherwise.
}
\examples{
\dontrun{
if(interactive()){
 L = list(orig = letters, subset = sample(letters, 5))
 head(mat01(L))
 }
}
}
