% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCPCounter_score.R
\name{MCPCounter_score}
\alias{MCPCounter_score}
\title{Run MCPCounter on a data.frame}
\usage{
MCPCounter_score(
  x,
  feat_types = c("HUGO_symbols", "ENTREZ_ID"),
  gene_pops = genes,
  out_digits = 2
)
}
\arguments{
\item{x}{data.frame with features in rows and samples in columns.}

\item{feat_types}{type of identifiers for expression features. Defaults to "HUGO_symbols" (Official gene symbols) and "ENTREZ_ID" (Entrez Gene ID).}

\item{out_digits}{integer indicating the number of decimal places (round). negative values are allowed.}

\item{gene_props}{data.frame of genes transcriptomic markers (HUGO symbols or ENTREZ_ID) and corresponding cell populations.}
}
\value{
data.frame of MCPCounter scores, where the first column is for each mixture.
}
\description{
Runs MCPCounter and produces a data.frame with abundance estimates.
}
