% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ncols}
\alias{ncols}
\title{<ncol> for many matrices}
\usage{
ncols(mats)
}
\arguments{
\item{m}{a list of matrices (or a single matrix)}
}
\value{
ncol for each matrix provided.
}
\description{
Returns the result of ncol for every matrix in a list
}
