% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nSigGenes.R
\name{nSigGenes}
\alias{nSigGenes}
\title{FUNCTION_TITLE}
\usage{
nSigGenes(
  group = NULL,
  m = NULL,
  p = c(0.01, 0.001, 1e-04),
  center.rows = TRUE,
  lfc = log2(2L),
  pmethod = "BH",
  alternative = "greater",
  simplify = TRUE
)
}
\arguments{
\item{group}{PARAM_DESCRIPTION, Default: NULL}

\item{m}{PARAM_DESCRIPTION, Default: NULL}

\item{p}{PARAM_DESCRIPTION, Default: c(0.01, 0.001, 1e-04)}

\item{center.rows}{PARAM_DESCRIPTION, Default: TRUE}

\item{lfc}{PARAM_DESCRIPTION, Default: log2(2L)}

\item{pmethod}{PARAM_DESCRIPTION, Default: 'BH'}

\item{alternative}{PARAM_DESCRIPTION, Default: 'greater'}

\item{simplify}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
