% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntop.R
\name{ntop}
\alias{ntop}
\title{Select top n elements}
\usage{
ntop(x, n, remove.lower = FALSE)
}
\arguments{
\item{x}{A vector or list.}

\item{n}{Number of elements to return.}

\item{remove.lower}{Remove vectors that have fewer than 'n' elements. Default: FALSE}
}
\value{
A vector or list of the same type as 'x', with a maximum of 'n' elements per vector.
}
\description{
ntop() returns the top 'n' elements of a vector or per vector in a list.
}
