% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohmyqc.R
\name{ohmyqc}
\alias{ohmyqc}
\title{Quality control statistics for single cell RNA seq data}
\usage{
ohmyqc(m, mito.genes = rownames(m)[startsWith(rownames(m), "MT-")])
}
\arguments{
\item{m}{expression matrix of genes by cells in non-log2 space e.g. CPM or TPM.}

\item{mito.genes}{character vector of mitochondrial gene names. Default: rownames(m)[startsWith(rownames(m), "MT-")]}
}
\value{
dataframe with columns: cell, detected_genes, housekeeping_expr, mito_Frac
}
\description{
Quality control statistics for single cell RNA seq data
}
\seealso{
\code{\link[plyr]{mapvalues}}
}
