% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohmySNNclusters.R
\name{ohmySNNclusters}
\alias{ohmySNNclusters}
\title{SNN Graph-Based Clustering}
\usage{
ohmySNNclusters(
  m,
  all.methods = FALSE,
  methods = c("walktrap", "louvain"),
  k = c(10, 30, 50),
  ndim = nrow(m),
  ...
)
}
\arguments{
\item{m}{matrix of features (genes/principal components) by observations (cells)}

\item{all.methods}{logical value. Methods: walktrap, louvain, infomap, fast_greedy, label_prop, leading_edge. See the igraph package for details. Default: FALSE}

\item{methods}{one or more clustering methods to use. If multiple, each will be run. Default: c("walktrap", "louvain")}

\item{k}{number of nearest neighbours to use in the graph construction. If multiple values are inputted, each will be run. Default: c(10, 30, 50)}

\item{ndim}{number of dimensions to use in the graph construction. If multiple values are inputted, each will be run. Default: nrow(m)}

\item{...}{other arguments for graph construction passed to scran::buildSNNGraph}
}
\value{
a dataframe
}
\description{
We first build a SNN graph, where each node is a cell connected to its neighbours in high dimensional space. Input multiple values for 'k' and 'ndim' -- the nearest neighbours and number of dimensions to use --- and the function will generate a graph for every combination. We then apply a clustering method to each of the graph(s) to identify clusters of cells. You can choose multiple clustering methods and the function will assign cells to clusters based on each method and for each graph. The clustering methods are taken from the igraph package. For a good and practical introduction of graph-based clustering, see: https://osca.bioconductor.org/clustering.html#clustering-graph
}
\details{
The function provides a friendly interface for trying several combinations of parameters for graph construction and cluster identification. The output is a dataframe with columns: id, k, ndim and one column per clustering method.
}
\seealso{
\code{\link[igraph]{cluster_walktrap}},\code{\link[igraph]{cluster_louvain}},\code{\link[igraph]{cluster_infomap}},\code{\link[igraph]{cluster_fast_greedy}},\code{\link[igraph]{cluster_label_prop}},\code{\link[igraph]{cluster_leading_eigen}}
 \code{\link[scran]{buildSNNGraph}}
 \code{\link[dplyr]{join}}
}
