% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohmytsne.R
\name{ohmytsne}
\alias{ohmytsne}
\title{A plot-friendly output for the Barnes-Hut implementation of t-SNE.}
\usage{
ohmytsne(m, perplexity = c(15, 30, 50), ndim = ncol(m), ...)
}
\arguments{
\item{m}{matrix; each row is an observation (eg. cell); each column is a feature (eg. gene / principal component)}

\item{perplexity}{numeric vector of perplexities to use. Default: c(15, 30, 50)}

\item{ndim}{numeric vector of dimensions to use. Default: ncol(m)}

\item{...}{other arguments passed to Rtsne::Rtsne}
}
\value{
a dataframe with columns: id, tSNE1, tSNE2, perplexity, ndim
}
\description{
A plot-friendly output for the Barnes-Hut implementation of t-SNE: test different perplexities and dimensions. See Rtsne::Rtsne for details on the algorithm.
}
\details{
Every combination of values for 'perplexity' and 'ndim' will be run if multiple are provided for each.
}
\seealso{
\code{\link[Rtsne]{Rtsne}}
 \code{\link[dplyr]{join}}
}
