% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\title{FUNCTION_TITLE}
\usage{
outliers(formula, data, mod = loess, n.sd = 1.5, ...)
}
\arguments{
\item{formula}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION}

\item{mod}{PARAM_DESCRIPTION, Default: loess}

\item{n.sd}{PARAM_DESCRIPTION, Default: 1.5}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
