% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairup.R
\name{pairup}
\alias{pairup}
\title{List combinations}
\usage{
pairup(..., by = c("intersect", "union", "c"))
}
\arguments{
\item{...}{Unnamed list(s). If only one list is provided, it will be combined (pairwise) with itself.}

\item{by}{Function to apply. Default: c("intersect", "union", "c")}
}
\value{
List with combined elements
}
\description{
Combine all pairs of list elements
}
\seealso{
\code{\link[tidyr]{expand_grid}}
 \code{\link[stats]{setNames}}
}
