% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal19-data.R
\docType{data}
\name{pal19}
\alias{pal19}
\title{pal19, 19 Pretty Discrete Colours}
\format{
A character vector.
}
\usage{
pal19
}
\description{
pal19, 19 Pretty Discrete Colours
}
\examples{
scales::show_col(pal19)
}
\keyword{datasets}
