% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal28-data.R
\docType{data}
\name{pal28}
\alias{pal28}
\title{pal28, 28 Pretty Discrete Colours}
\format{
A character vector.
}
\usage{
pal28
}
\description{
pal28, 28 Pretty Discrete Colours
}
\examples{
scales::show_col(pal28)
}
\keyword{datasets}
