% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palBuRd-data.R
\docType{data}
\name{palBuRd}
\alias{palBuRd}
\title{Divergent Heatmap Colours from Blue to Red}
\format{
A character vector of 11 heatmap colours from blue to red, with the middle value as whote
}
\usage{
palBuRd
}
\description{
Divergent Heatmap Colours from Blue to Red
}
\examples{
scales::show_col(palBuRd)
}
\keyword{datasets}
