% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palG34R-data.R
\docType{data}
\name{palG34R}
\alias{palG34R}
\title{palG34R, a palette of reds and blues for cases within two groups}
\format{
A character vector.
}
\usage{
data(palG34R)
}
\description{
palG34R, a palette of reds and blues for cases within two groups
}
\examples{
data(palG34R)
scales::show_col(palG34R)
}
\keyword{datasets}
