% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{FUNCTION_TITLE}
\usage{
pca(
  m,
  npcs = 100,
  retx = TRUE,
  center = TRUE,
  grouped.center = NULL,
  scale = FALSE
)
}
\arguments{
\item{m}{PARAM_DESCRIPTION}

\item{npcs}{PARAM_DESCRIPTION, Default: 100}

\item{retx}{PARAM_DESCRIPTION, Default: TRUE}

\item{center}{PARAM_DESCRIPTION, Default: TRUE}

\item{grouped.center}{PARAM_DESCRIPTION, Default: NULL}

\item{scale}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[irlba]{prcomp_irlba}}
}
