% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hierarchy.R
\name{plot_hierarchy}
\alias{plot_hierarchy}
\title{Plot State Hierarchy (Quadrant plot)}
\usage{
plot_hierarchy(
  X,
  quadrant.names = c("bl", "br", "tl", "tr"),
  main = NULL,
  xlab = "Relative meta-module score [log2(|SC1-SC2|+1)]",
  ylab = "Relative meta-module score [log2(|SC1-SC2|+1)]",
  groups = NULL,
  group.cols = NULL,
  legend = T,
  legend.pos = "bottom",
  legend.horiz = T
)
}
\arguments{
\item{X}{two-column matrix or dataframe with x and y coordinates.}

\item{quadrant.names}{character vector to label quadrants. Default: c("bl", "br", "tl", "tr")}

\item{main}{plot title. Default: NULL}

\item{xlab}{x-axis title. Default: 'Relative meta-module score [log2(|SC1-SC2|+1)]'}

\item{ylab}{y-axis title. Default: 'Relative meta-module score [log2(|SC1-SC2|+1)]'}

\item{groups}{a named character vector, where the vector corresponds to rownames in X and names are the groups to colour by. Default: NULL}

\item{group.cols}{named character vector of colours to use for each group. Names should match names in groups. Default: NULL}

\item{legend}{logical to draw legend if groups are provided. Default: T}

\item{legend.pos}{legend position. Default: 'bottom'}

\item{legend.horiz}{logical indicating whether legend should be horizontally drawn. Default: T}
}
\value{
base plot
}
\description{
Plot state hierarchy
}
