% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{programs}
\alias{programs}
\title{}
\usage{
programs(
  m,
  nsig1 = 50,
  nsig2 = 10,
  jaccard = 0.7,
  p = 0.01,
  lfc = log2(2),
  pmethod = "BH"
)
}
\arguments{
\item{m}{PARAM_DESCRIPTION}

\item{nsig1}{PARAM_DESCRIPTION, Default: 50}

\item{nsig2}{PARAM_DESCRIPTION, Default: 10}

\item{jaccard}{PARAM_DESCRIPTION, Default: 0.7}

\item{p}{PARAM_DESCRIPTION, Default: 0.01}

\item{lfc}{PARAM_DESCRIPTION, Default: log2(2)}

\item{pmethod}{PARAM_DESCRIPTION, Default: 'BH'}
}
\value{
********

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
**********

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
