% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_rowmeans.R
\name{range_rowmeans}
\alias{range_rowmeans}
\title{Range of the rowMeans of a matrix}
\usage{
range_rowmeans(mat)
}
\arguments{
\item{mat}{a matrix}
}
\value{
the range of the rowMeans of the matrix provided.
}
\description{
Returns the range over the rowMeans of a matrix
}
