% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{reorder}
\alias{reorder}
\title{Reorder matrix columns and rows}
\usage{
reorder(m, col = T, row = T)
}
\arguments{
\item{m}{a matrix}

\item{col}{logical indicating whether to reorder columns. Default: T}

\item{row}{logical indicating whether to reorder rows. Default: T}
}
\value{
ordered matrix
}
\description{
Reorder matrix columns and rows by hierarchical clustering of pairwise pearson correlation values
}
