% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_exclude.R
\name{row_exclude}
\alias{row_exclude}
\title{Drop rows whose names match a pattern}
\usage{
row_exclude(m, pattern)
}
\arguments{
\item{m}{Input matrix or data frame.}

\item{pattern}{Pattern to look for. The default interpretation is a regular expression, as described in stringi::stringi-search-regex. Control options with 'regex()'.}
}
\value{
A matrix.
}
\description{
row_exclude() is a wrapper around m[!str_detect(rownames(m), pattern), ]. Equivalent to row_subset() with negate = TRUE.
}
\details{
Vectorised over rownames('m') and 'pattern'
}
