% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_subset.R
\name{row_subset}
\alias{row_subset}
\title{Keep rows whose names match a pattern}
\usage{
row_subset(m, pattern, negate = FALSE)
}
\arguments{
\item{m}{Input matrix or data frame.}

\item{pattern}{Pattern to look for. The default interpretation is a regular expression, as described in stringi::stringi-search-regex. Control options with 'regex()'.}

\item{negate}{If 'TRUE', return non-matching rows.}
}
\value{
A matrix.
}
\description{
row_subset() is a wrapper around m[, str_detect(rownames(m), pattern)]
}
\details{
Vectorised over rownames('m') and 'pattern'
}
