% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-ops.R
\name{rowcenter}
\alias{rowcenter}
\title{Center a matrix row-wise}
\usage{
rowcenter(m, by = "mean")
}
\arguments{
\item{m}{a matrix or Matrix}

\item{by}{either "mean", "median" or a numeric vector of length equal to the number of rows of ‘m’. Default: "mean"}
}
\value{
row-centered matrix
}
\description{
Center a matrix row-wise
}
