% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_strict.R
\name{sample_strict}
\alias{sample_strict}
\title{Assign Rows and Sample}
\usage{
sample_strict(mat, size, min = 1, diff = 0.5, replace = F)
}
\arguments{
\item{mat}{a matrix}

\item{size}{sample size}

\item{min}{minimum value required for a row to be assigned to a column. Default: 1}

\item{diff}{minimum difference in value to the 'next-best' row that is required for a row to be assigned to a column. Default: 0.5}

\item{replace}{boolean value indicating whether sample replacement is allowed. Will be set to TRUE with warning if the sample is smaller than the sample size. Default: F}
}
\value{
a list of sampled rows for each column, each elemnent a character vector of size <size>
}
\description{
Assign rows to columns of a matrix with maxcol_strict and subsequently sample from each column's group.
}
