% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setRowNames.R
\name{setRowNames}
\alias{setRowNames}
\title{Set the rownames in an object}
\usage{
setRowNames(object = nm, nm)
}
\arguments{
\item{object}{an object for which a ‘rownames’ attribute will be meaningful. Default: nm}

\item{nm}{a character vector of rownames to assign to the object}
}
\value{
An object of the same sort as ‘object’ with the new rownames
}
\description{
This is a convenience function that sets the rownames on an object, in the same manner as stats::setNames
}
