% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signatures_Gavish2023.R
\docType{data}
\name{Signatures_Gavish2023}
\alias{Signatures_Gavish2023}
\title{Hallmarks of transcriptional intratumour heterogeneity across a thousand tumours}
\format{
list of character vectors
}
\usage{
Signatures_Gavish2023
}
\description{
Gavish, A., Tyler, M., Greenwald, A.C. et al. Hallmarks of transcriptional intratumour heterogeneity across a thousand tumours. Nature 618, 598–606 (2023). https://doi.org/10.1038/s41586-023-06130-4
}
\keyword{datasets}
