% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpsons-data.R
\docType{data}
\name{simpsons}
\alias{simpsons}
\title{simpsons, 16 Discrete Colours}
\format{
A character vector.
}
\usage{
simpsons
}
\description{
simpsons, 16 Discrete Colours
}
\examples{
scales::show_col(simpsons)
}
\keyword{palettes}
