% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_by_sample.R
\name{split_by_match}
\alias{split_by_match}
\title{Split Character Vector by Pattern Matching}
\usage{
split_by_match(x, pattern)
}
\arguments{
\item{x}{character vector}

\item{pattern}{character vector with pattern(s) to match}
}
\value{
list of character vectors with patterns as list names
}
\description{
Split character vector into subsets according to detected pattern matches.
}
\seealso{
\code{\link[stringr]{str_detect}}
}
