% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_by_sample_names.R
\name{split_by_sample_names}
\alias{split_by_sample_names}
\title{Split Character Vector of Cell IDs by Sample Names}
\usage{
split_by_sample_names(x, samples = NULL, ...)
}
\arguments{
\item{x}{character vector of cell IDs}

\item{samples}{a character vector of sample names, or NULL. If NULL, uses scalop::unique_sample_names to find samples.}

\item{...}{other arguments passed to scalop::unique_sample_names.}
}
\value{
list of character vectors with samples as list names
}
\description{
Split character vector of cell ids into subsets according to detected sample names.
}
\seealso{
\code{\link[stringr]{str_detect}}
}
