% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-ops.R
\name{split_matrix}
\alias{split_matrix}
\title{Split a matrix by a variable}
\usage{
split_matrix(m, by)
}
\arguments{
\item{m}{A  matrix}

\item{by}{A variable to split by}
}
\value{
A list of matrices split by a given variable.
}
\description{
Splits a matrix by a given variable
}
