% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substri.R
\name{substri}
\alias{substri}
\title{Extract substrings between delimiters by position}
\usage{
substri(x, sep = "\\\\.|-|_", pos = 1, max.nchar = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{A character vector.}

\item{sep}{The pattern to split each element of 'x' by. Default: '\\.|-|_'}

\item{pos}{The substring index(es) to be returned. Default: 1}

\item{max.nchar}{'NULL', or a maximum number of characters that the substrings should be trimmed to contain. Default: NULL}

\item{na.rm}{Remove NA substring positions. Default: TRUE}
}
\value{
A character vector containing the extracted substrings.
}
\description{
substri() splits each element in 'x' into substrings and returns the desired substring(s). 'sep' defines the delimiters separating the strings into substrings, such that the input data between the matches become the fields themselves. 'pos' defines the substring index(es) to be returned.
}
\examples{
\dontrun{
x = c('dog.cat.mat', 'dog_frog')
substri(x)
substri(x, pos = 2)
substri(x, pos = c(1,3))
substri(x, pos = c(1,3), na.rm = FALSE)
substri(x, sep = "-")
}
}
\seealso{
\code{\link[stringr]{str_split}},\code{\link[stringr]{str_replace_all}}
}
