% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{taill}
\alias{taill}
\title{Left-hand tail a matrix or dataframe}
\usage{
taill(m, i = 6L)
}
\arguments{
\item{m}{a matrix or dataframe}

\item{i}{number of columns to print}
}
\description{
Left-hand tail a matrix or dataframe
}
