% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tailr}
\alias{tailr}
\title{Right-hand tail a matrix or dataframe}
\usage{
tailr(m, i = 6L)
}
\arguments{
\item{m}{a matrix or dataframe}

\item{i}{number of columns to print}
}
\description{
Right-hand tail a matrix or dataframe
}
