% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_scalop.R
\name{theme_scalop}
\alias{theme_scalop}
\title{FUNCTION_TITLE}
\usage{
theme_scalop(
  title.size = 16,
  text.size = 14,
  axis.title.size = title.size,
  legend.title.size = title.size,
  strip.text.size = title.size,
  axis.text.size = text.size,
  legend.text.size = text.size,
  legend.position = "bottom",
  legend.justification = "right",
  text.colour = "black",
  ...
)
}
\arguments{
\item{title.size}{PARAM_DESCRIPTION, Default: 16}

\item{text.size}{PARAM_DESCRIPTION, Default: 14}

\item{axis.title.size}{PARAM_DESCRIPTION, Default: title.size}

\item{legend.title.size}{PARAM_DESCRIPTION, Default: title.size}

\item{strip.text.size}{PARAM_DESCRIPTION, Default: title.size}

\item{axis.text.size}{PARAM_DESCRIPTION, Default: text.size}

\item{legend.text.size}{PARAM_DESCRIPTION, Default: text.size}

\item{legend.position}{PARAM_DESCRIPTION, Default: 'bottom'}

\item{legend.justification}{PARAM_DESCRIPTION, Default: 'right'}

\item{text.colour}{PARAM_DESCRIPTION, Default: 'black'}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{ggtheme}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}}
 \code{\link[grid]{unit}}
}
