% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{ttest}
\alias{ttest}
\title{t-tests}
\usage{
ttest(x, y, ...)
}
\arguments{
\item{x}{matrix or character vector with matrix column names}

\item{y}{matrix to compare against or matrix with x and y column names}

\item{...}{adjust.method and cutoff}
}
\value{
t-tests
}
\description{
t-tests
}
