% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_sample_names.R
\name{unique_sample_names}
\alias{unique_sample_names}
\title{Get unique sample names}
\usage{
unique_sample_names(x, sep = "-|_", pos = 1, max.nchar = NULL, replace = NULL)
}
\arguments{
\item{x}{character vector of cell ids}

\item{sep}{separator Default: '-|_'}

\item{pos}{position of sample name given separator. First position is 1. Default: 1}

\item{max.nchar}{maximum number of characters to take as sample name (starting from the first). Default: NULL}

\item{replace}{a list of character vectors to replace. Takes the form list(c(old, new), c(old, new)). Default: NULL}
}
\value{
character vector of unique sample names
}
\description{
Extract unique sample names from cell ids
}
