% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unlist.R
\name{Unlist}
\alias{Unlist}
\title{Unlist, keeping original list or element names}
\usage{
Unlist(L, nested.names = FALSE)
}
\arguments{
\item{L}{list to flatten}

\item{nested.names}{logical; keep nested list names rather than expanding list names. Default: FALSE}
}
\value{
a vector of the same length as the combined lengths of list elements. Names will either be the list names replicated, or, if nested.names is TRUE, the original list element names will be kept.
}
\description{
Unlist, keeping original list or element names
}
\seealso{
\code{\link[stats]{setNames}}
}
