% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vannomap.R
\name{vannomap}
\alias{vannomap}
\title{One-bar annotation ggplot}
\usage{
vannomap(
  X,
  y.name = NULL,
  x.name = NULL,
  angle = F,
  num = T,
  flip = F,
  hide.legend = T,
  get.legend = F,
  legend.pos = "top",
  ratio = 33,
  tile.col = "black",
  tile.size = 0.01,
  mar = 0.015,
  cols = NULL,
  limits = NULL,
  nrow = 1,
  patchwork = T,
  heights = c(9, 1),
  guides = "collect",
  ...
)
}
\arguments{
\item{X}{named character vector of observations and groups (names); or a dataframe with first two columns as observations and groups.}

\item{angle}{logical indicating whether to angle title if flip = T. Default: T}

\item{num}{enumerate observations on x axis? Default: T}

\item{flip}{observations along y axis instead of x? Default: F}

\item{hide.legend}{logical. Default: T}

\item{legend.pos}{character string. See legend.position in ggplot2::theme for details. Default: 'top'}

\item{ratio}{aspect ratio. If flip = T, ratio = 1/ratio.. Default: 0.03}

\item{mar}{see plot.margin in ggplot2::themes for details. Default: 0.015}

\item{cols}{custom colour palette. Option to name colours by groups to choose individual group colours. Default: NULL}

\item{title}{vannomap title. Default: NULL}

\item{x.title}{x axis (observations) title (e.g. 'cells'). Default: NULL}

\item{pal}{numeric value specifying different qualitative brewer palettes. See ggplot2::scale_fill_brewer for details. Default: 1}

\item{breaks}{set to NULL to remove x axis tick marks. Default: ggplot2::waiver()}

\item{cols.order}{character vector with groups. Changing the order changes the assignment of colours to groups. Default: names(cols)}
}
\value{
ggplot
}
\description{
one-bar annotation ggplot
}
\seealso{
\code{\link[ggplot2]{waiver}},\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{scale_continuous}},\code{\link[ggplot2]{scale_x_discrete}},\code{\link[ggplot2]{geom_raster}},\code{\link[ggplot2]{ggtheme}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}},\code{\link[ggplot2]{scale_colour_brewer}},\code{\link[ggplot2]{scale_manual}},\code{\link[ggplot2]{coord_flip}}
 \code{\link[ggpubr]{get_legend}},\code{\link[ggpubr]{rremove}}
}
