% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_genes.R
\name{which_annotation}
\alias{which_annotation}
\title{FUNCTION_TITLE}
\usage{
which_annotation(keys, aliasesBeforeSymbols = FALSE, priority = NULL)
}
\arguments{
\item{keys}{PARAM_DESCRIPTION}

\item{aliasesBeforeSymbols}{PARAM_DESCRIPTION, Default: FALSE}

\item{priority}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
