% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxtest.R
\name{wilcoxtest}
\alias{wilcoxtest}
\title{wilcoxon-tests}
\usage{
wilcoxtest(x, y, ...)
}
\arguments{
\item{x}{matrix or character vector with matrix column names}

\item{y}{matrix to compare against or matrix with x and y column names}

\item{...}{adjust.method and cutoff}
}
\value{
wilcoxon-tests
}
\description{
wilcoxon-tests
}
