% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permuteSigScores.R
\name{binom.test_matrix}
\alias{binom.test_matrix}
\title{FUNCTION_TITLE}
\usage{
binom.test_matrix(m, adjust.method = "fdr", MARGIN = 1)
}
\arguments{
\item{m}{PARAM_DESCRIPTION}

\item{adjust.method}{PARAM_DESCRIPTION, Default: 'fdr'}

\item{MARGIN}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{p.adjust}}
}
