% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enricher.R, R/old-enricher.R
\name{enricher}
\alias{enricher}
\title{Functional enrichments}
\usage{
enricher(
  my_gene_sets,
  my_universe,
  p.adj = 0.01,
  term2gene = NULL,
  category = c("C2", "C5"),
  subcategory = NULL
)

enricher(
  my_gene_sets,
  my_universe,
  p.adj = 0.01,
  term2gene = NULL,
  category = c("C2", "C5"),
  subcategory = NULL
)
}
\arguments{
\item{my_gene_sets}{PARAM_DESCRIPTION}

\item{my_universe}{PARAM_DESCRIPTION}

\item{p.adj}{PARAM_DESCRIPTION, Default: 0.01}

\item{term2gene}{PARAM_DESCRIPTION, Default: NULL}

\item{category}{PARAM_DESCRIPTION, Default: c("C2", "C5")}

\item{subcategory}{PARAM_DESCRIPTION, Default: NULL}

\item{test_gene_sets}{(list of) character vector gene IDs}

\item{ref_gene_sets}{(list of) character vector gene IDs}

\item{universe}{all genes considered. If NULL, uses all genes in test_gene_sets. Default: NULL}

\item{minGSSize}{minimum gene set size, Default: 10}

\item{maxGSSize}{maximum gene set size, Default: 500}

\item{pAdjustMethod}{p-value adjustment method, Default: 'BH'}

\item{pvalueCutoff}{p-value cutoff, Default: 0.05}

\item{qvalueCutoff}{FDR cutoff, Default: 0.2}
}
\value{
dataframe with enrichment statistics.

OUTPUT_DESCRIPTION
}
\description{
Functional enrichments by hypergeometric test

FUNCTION_DESCRIPTION
}
\details{
enriched ref_gene_sets per test_gene_set are returned. test_gene_set names are specified in the name column and ref_gene_set name are specified in the ID column.

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[limma]{alias2Symbol}}
 \code{\link[clusterProfiler]{enricher}}

\code{\link[msigdbr]{msigdbr}}
 \code{\link[clusterProfiler]{enricher}}
}
