% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enricher.R
\name{enricher}
\alias{enricher}
\title{Functional enrichments}
\usage{
enricher(
  test_gene_sets,
  ref_gene_sets,
  universe = NULL,
  minGSSize = 10,
  maxGSSize = 500,
  pAdjustMethod = "BH",
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.2
)
}
\arguments{
\item{test_gene_sets}{(list of) character vector gene IDs}

\item{ref_gene_sets}{(list of) character vector gene IDs}

\item{universe}{all genes considered. If NULL, uses all genes in test_gene_sets. Default: NULL}

\item{minGSSize}{minimum gene set size, Default: 10}

\item{maxGSSize}{maximum gene set size, Default: 500}

\item{pAdjustMethod}{p-value adjustment method, Default: 'BH'}

\item{pvalueCutoff}{p-value cutoff, Default: 0.05}

\item{qvalueCutoff}{FDR cutoff, Default: 0.2}
}
\value{
dataframe with enrichment statistics.
}
\description{
Functional enrichments by hypergeometric test
}
\details{
enriched ref_gene_sets per test_gene_set are returned. test_gene_set names are specified in the name column and ref_gene_set name are specified in the ID column.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[limma]{alias2Symbol}}
 \code{\link[clusterProfiler]{enricher}}
}
