% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sample_names.R
\name{extract_sample_names}
\alias{extract_sample_names}
\title{Extract Sample Names}
\usage{
extract_sample_names(
  x,
  samples = NULL,
  sep = "-|_",
  pos = 1,
  max.nchar = NULL,
  replace = NULL
)
}
\arguments{
\item{x}{character vector of strings to extract sample names from.}

\item{samples}{character vector of unique sample names to extract from x. Default: NULL}

\item{sep}{separator. Relevant only if samples = NULL. Default: "-|_|\\."}

\item{pos}{position of sample name given separator. First position is 1. Relevant only if samples = NULL. Default: 1}

\item{max.nchar}{maximum number of characters to take as sample name (starting from the first). Relevant only if samples = NULL. Default: NULL}

\item{replace}{a list of character vectors to replace. Takes the form list(c(old, new), c(old, new)). Relevant only if samples = NULL. Default: NULL}
}
\value{
character vector of sample names of the same length as x.
}
\description{
Convenient wrapper function for `scalop::extract_substring`. Extract sample names from character vector. Sample names can be provided or, if not, then are found with `scalop::unique_sample_names`.
}
