% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvolcano.R
\name{ggvolcano}
\alias{ggvolcano}
\title{Volcano plot}
\usage{
ggvolcano(
  m = NULL,
  group = NULL,
  group2 = NULL,
  dea.res = NULL,
  pmethod = "BH",
  alternative = NULL,
  logp = 2,
  lfc = 1,
  n = 10,
  selected = NULL,
  add.to.selected = NULL,
  selected.col = "red",
  xlab = "Log2 Fold Change",
  ylab = "-Log10 P-value (adjusted)",
  symmetric = TRUE,
  xintercept = 0,
  segment.colour = "grey30",
  text.size = 14,
  label.size = 5,
  legend.text.size = 12,
  ymax = NULL,
  xlim = NULL
)
}
\arguments{
\item{m}{A matrix of features (e.g. genes) by observations (e.g. cells). Default: NULL}

\item{group}{The group (of observations) to test. This should be a character vector or a two-level factor in which the first level corresponds to the group. Default: NULL}

\item{group2}{A second group to test against. This should be a character vector. If 'group2=NULL', defaults to all remaining observations in the matrix 'm'. Default: NULL}

\item{dea.res}{The output of scalop::dea(..., return.val = 'df'). If provided, the function will not perform DEA and skip straight to plotting. Default: NULL}

\item{pmethod}{The correction method for multiple testing, or 'none' if no correction is desired. See stats::p.adjust.methods for options. Default: 'BH'}

\item{alternative}{The alternative hypothesis. If 'alternative=NULL' and 'group2' was provided, defaults to 'two-sided'. Else defaults to 'greater'. Default: NULL}

\item{logp}{The minimum -log10(P) value at which differential expression is considered significant. Default: 2}

\item{lfc}{The minimum log2 foldchange value at which differential expression is considered signficant. Default: 1}

\item{n}{The number of top features to display. Default: 10}

\item{selected}{Default: NULL}

\item{add.to.selected}{PARAM_DESCRIPTION, Default: NULL}

\item{selected.col}{PARAM_DESCRIPTION, Default: 'red'}

\item{xlab}{X-axis title. Default: 'Log2 Fold Change'}

\item{ylab}{Y-axis title. Default: '-Log10 P-value (adjusted)'}

\item{symmetric}{PARAM_DESCRIPTION, Default: TRUE}

\item{xintercept}{One or more numeric values to be added as vertical delimiters to the plot. Default: 0}

\item{text.size}{Axes text size. Default: 14}

\item{label.size}{Feature label text sizes. Default: 5}

\item{legend.text.size}{Legend text size. Default: 12}

\item{ymax}{Y-axis upper limit. If 'NULL', defaults to the maximum -log10(P) value. Setting 'ymax' can be useful to 'zoom in' on the y-axis, making feature points/labels below 'ymax' easier to see. Default: NULL}

\item{xlim}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
A list containing:
\itemize{
\item {data:} {A data frame; the differential expression results used in the plotting.}
\item {G:} {A ggplot2 object; the volcano plot.}
}
}
\description{
Perform differential expression and create a volcano plot.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{pull}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{scale_continuous}},\code{\link[ggplot2]{ggtheme}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{scale_manual}},\code{\link[ggplot2]{geom_abline}}
 \code{\link[ggrepel]{geom_text_repel}}
}
