% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca.R
\name{hca}
\alias{hca}
\alias{hca_cor}
\alias{hca_dist}
\alias{hca_tree}
\alias{hca_order}
\alias{hca_groups}
\title{FUNCTION_TITLE}
\usage{
hca(x, cor.method = cor.methods, dist.method = dist.methods,
  cluster.method = cluster.methods, h = NULL, k = NULL,
  groups.minsize = 5, groups.maxsize = 0.5)

hca_cor(x, return.steps = F, ...)

hca_dist(x, return.steps = F, ...)

hca_tree(x, return.steps = F, ...)

hca_order(x, return.steps = F, ...)

hca_groups(x, return.steps = F, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{cor.method}{PARAM_DESCRIPTION, Default: cor.methods}

\item{dist.method}{PARAM_DESCRIPTION, Default: dist.methods}

\item{cluster.method}{PARAM_DESCRIPTION, Default: cluster.methods}

\item{h}{PARAM_DESCRIPTION, Default: NULL}

\item{k}{PARAM_DESCRIPTION, Default: NULL}

\item{groups.minsize}{PARAM_DESCRIPTION, Default: 5}

\item{groups.maxsize}{PARAM_DESCRIPTION, Default: 0.5}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
