% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Markers_Normal.R
\docType{data}
\name{Markers_Normal}
\alias{Markers_Normal}
\title{Marker gene sets of Normal Cell Types and Pan-Cancer Programs}
\format{
A list of character vectors (i.e. sets of gene per cell type / program)
}
\source{
Microglia signature: \url{doi:10.1016/j.cell.2019.11.010}; Housekeeping signature: \url{doi:10.1126/science.aad0501}.
}
\usage{
Markers_Normal
}
\description{
Marker genes for different normal cell types and some pan-cancer expression programs.
}
\keyword{datasets}
