% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaprograms.R
\name{metaprograms}
\alias{metaprograms}
\title{Compute Metaprograms}
\usage{
metaprograms(
  programs,
  profiles,
  m,
  metaclusters,
  samples = NULL,
  freq.cutoff = 3,
  order.priority = c("freq", "lfc"),
  return.data = FALSE
)
}
\arguments{
\item{programs}{list of differentially expressed genes in each group/cell cluster}

\item{profiles}{list of differential expression profiles for all genes in each group/cell cluster}

\item{m}{matrix of genes by cells}

\item{metaclusters}{list of program metaclusters}

\item{samples}{character vector of unique sample names. Default: readRDS("samples_2020-03-01.rds")}

\item{order.priority}{order by 'freq' first or 'lfc' first. Default: 'freq'}

\item{freq}{minimum number of metacluster-specific sample-specific programs a gene must appear in. Default: 3}
}
\value{
if return.data = F, just the list of metaprograms. Else, the list of metaprograms, and the dataframe of metacluster-, sample- and gene- specific LFC and freq values.
}
\description{
programs and profiles are the lists of gene sets and DE profiles that were derived from intra-tumour cell clusters. These are integrated, according to the metaclusters clustering, into a final list of metaprograms; one per metacluster.
}
\details{
DETAILS
}
\seealso{
\code{\link[reshape2]{melt}}
 \code{\link[dplyr]{select}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{join}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{distinct}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{arrange}}
}
