% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-ops.R
\name{nrows}
\alias{nrows}
\title{<nrow> for many matrices}
\usage{
nrows(mats)
}
\arguments{
\item{mats}{a list of matrices (or a single matrix)}
}
\value{
nrow for each matrix provided.
}
\description{
Returns the result of nrow for every matrix in a list
}
