% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permuteSigScores.R
\name{permute_matrix}
\alias{permute_matrix}
\title{FUNCTION_TITLE}
\usage{
permute_matrix(m, MARGIN = 1)
}
\arguments{
\item{m}{PARAM_DESCRIPTION}

\item{MARGIN}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[scalop]{setColNames}}, \code{\link[scalop]{setRowNames}}
}
