% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permuteSigScores.R
\name{permuteSigScores}
\alias{permuteSigScores}
\title{FUNCTION_TITLE}
\usage{
permuteSigScores(
  m,
  sigs,
  center.fun = NULL,
  center.rows = TRUE,
  expr.bin.m = NULL,
  N = 50,
  alternative = c("greater", "two.sided", "less"),
  ...
)
}
\arguments{
\item{m}{PARAM_DESCRIPTION}

\item{sigs}{PARAM_DESCRIPTION}

\item{center.fun}{PARAM_DESCRIPTION, Default: NULL}

\item{center.rows}{PARAM_DESCRIPTION, Default: TRUE}

\item{expr.bin.m}{PARAM_DESCRIPTION, Default: NULL}

\item{N}{PARAM_DESCRIPTION, Default: 50}

\item{alternative}{PARAM_DESCRIPTION, Default: c("greater", "two.sided", "less")}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[scalop]{sigScores}}, \code{\link[scalop]{setRowNames}}, \code{\link[scalop]{setColNames}}
 \code{\link[tibble]{rownames}}
 \code{\link[reshape2]{melt}}
 \code{\link[dplyr]{rename}}, \code{\link[dplyr]{mutate}}
}
