% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowdetected.R
\name{rowdetected}
\alias{rowdetected}
\title{Number of non-zero values per row}
\usage{
rowdetected(m)
}
\arguments{
\item{m}{matrix}
}
\value{
numeric vector
}
\description{
Number of non-zero values per row
}
\seealso{
\code{\link[matrixStats]{rowCounts}}
 \code{\link[stats]{setNames}}
}
