% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Score matrix columns for groups of rows}
\usage{
score(
  mat,
  groups,
  binmat = NULL,
  bins = NULL,
  controls = NULL,
  bin.control = F,
  center = F,
  nbin = 30,
  n = 100,
  replace = F
)
}
\description{
This function has been deprecated and replaced with scalop::sigScores. Please see `?scalop::sigScores` for details.
}
