% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_by_sample.R
\name{split_by_sample}
\alias{split_by_sample}
\title{Split Character Vector of Cell IDs by Sample Names}
\usage{
split_by_sample(x, samples = NULL, ...)
}
\arguments{
\item{x}{character vector of cell IDs}

\item{samples}{a character vector of sample names, or NULL. If NULL, uses scalop::extract_sample_names to find samples.}

\item{...}{other arguments passed to scalop::extract_sample_names.}
}
\value{

}
\description{
Split character vector of cell ids into subsets according to detected sample names.
}
\seealso{
\code{\link[stringr]{str_detect}}
}
