% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_by_sample_names.R
\name{split_by_sample_names}
\alias{split_by_sample_names}
\title{Split Character Vector of Cell IDs by Sample Names}
\usage{
split_by_sample_names(
  x,
  samples = NULL,
  sep = "-|_",
  pos = 1,
  max.nchar = NULL,
  replace = NULL
)
}
\arguments{
\item{x}{character vector of cell IDs}

\item{samples}{a character vector of sample names, or NULL. If NULL, uses scalop::unique_sample_names to find samples.}

\item{sep}{ignored if <samples> is provided. separator Default: '-|_'}

\item{pos}{ignored if <samples> is provided. position of sample name given separator. First position is 1. Default: 1}

\item{max.nchar}{ignored if <samples> is provided. maximum number of characters to take as sample name (starting from the first). Default: NULL}

\item{replace}{ignored if <samples> is provided. a list of character vectors to replace. Takes the form list(c(old, new), c(old, new)). Default: NULL}
}
\value{
list of character vectors with samples as list names
}
\description{
Split character vector of cell ids into subsets according to detected sample names.
}
\seealso{
\code{\link[stringr]{str_detect}}
}
